requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
openpyxl>=3.1.0  # For Excel file support
xlrd>=2.0.1      # For legacy Excel files